internal/gocommand: stop copying stdout as soon as ctx is done
If ctx is done, that means (by definition) that the caller is no
longer interested in the output of the command. Since 'go list'
output can be quite long, we shouldn't waste time copying it
if we no longer care about it.

If the command doesn't shut down on its own in response to the initial
Interrupt signal, or if it triggers a bug in the kernel leading to a
deadlock, or I/O is otherwise extremely slow (perhaps due to
swapping), the savings from this earlier shutdown could be
significant. Even if not, the impact (or lack thereof) of this change
could tell us more about what's going on.

For golang/go#54461.

Change-Id: I0df067cb77496dacd29497f0995da6fc1a4a37f1
Reviewed-on: https://go-review.googlesource.com/c/tools/+/484741
gopls-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
Auto-Submit: Bryan Mills <bcmills@google.com>
Run-TryBot: Bryan Mills <bcmills@google.com>
Commit-Queue: Bryan Mills <bcmills@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>